ai_scraper_framework/requirements.txt
PyYAML>=5.0
playwright>=1.30
beautifulsoup4>=4.9.0
fastapi>=0.70.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0
python-multipart
httpx>=0.20.0
pytest>=6.0.0
ultralytics>=8.0.0
opencv-python>=4.0.0
numpy>=1.20.0
Pillow>=9.0.0
pytesseract>=0.3.0
spacy>=3.0.0
